backend/requirements.txt
flask
langchain
chromadb
pdfplumber
python-dotenv
sentence-transformers
openai
fastapi

